+++ ios-native-objectivec/AYPrinterDemo/AYPrinterDemo/AYSDK.framework/Headers/AYSDK.h
@@ -16,12 +16,14 @@
 #import <AYSDK/AYTsplCommand.h>
 #import <AYSDK/TTypes.h>
 #import <AYSDK/AYWifiCommand.h>
+#import <AYSDK/AYRGBCommand.h>
 #import <AYSDK/AYOtaHelper.h>
 #import <AYSDK/NSString+AY.h>
 #import <AYSDK/NSData+AY.h>
 #import <AYSDK/SdkInfo.h>
 #import <AYSDK/AYImageEnhance.h>
 #import <AYSDK/AYNetHelper.h>
+#import <AYSDK/AYOtaHelperCPCL.h>
 
 //! Project version number for AYSDK.
 FOUNDATION_EXPORT double AYSDKVersionNumber;